refactor unit test
