+++ app/src/main/java/br/com/concrete/tentacle/data/repositories/LoginRepository.kt
@@ -1,4 +1,4 @@
-package br.com.concrete.tentacle.repositories
+package br.com.concrete.tentacle.data.repositories
 
 import br.com.concrete.tentacle.data.models.BaseModel
 import br.com.concrete.tentacle.data.models.RequestLogin
