+++ app/src/main/java/br/com/concrete/tentacle/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package br.com.concrete.tentacle.di
 
-import br.com.concrete.tentacle.repositories.LoginRepository
-import br.com.concrete.tentacle.repositories.UserRepository
+import br.com.concrete.tentacle.data.repositories.LoginRepository
+import br.com.concrete.tentacle.data.repositories.UserRepository
 import org.koin.dsl.module.module
 
 val repositoryModule = module {
