--- app/src/test/java/br/com/concrete/tentacle/ExampleUnitTest.kt
@@ -1,18 +1,0 @@
-package br.com.concrete.tentacle.tentacle
-
-import org.junit.Assert.assertEquals
-import org.junit.Ignore
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@Ignore
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
